pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-security-alert-investigation"
version = "1.0.0"
description = "Security Alert Investigation environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn>=0.30,<1",
    "pydantic>=2.8,<3",
    "httpx>=0.27,<1",
    "openai>=2.7.2,<3",
    "openenv-core[core]>=0.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "security_alert_investigation.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["security_alert_investigation", "security_alert_investigation.server"]
package-dir = { "security_alert_investigation" = ".", "security_alert_investigation.server" = "server" }

[tool.setuptools.package-data]
security_alert_investigation = ["data/*.json"]
